refactor(mcp-servers): extract TransportAdvancedFields and warn on bad preset headers
Deduplicate HttpForm headers/SSE paths and log when non-object headers overlays are dropped during preset merge.
